increase broadcast buffer size to 256
We can easily transmit 100 canvas events per second, so keep this buffer big to avoid dropping clients.
Also, this is still not _that_ big, a 256 element broadcast queue is probably only 32k.
Doubling that for the sum of inbound queues = 64k * 16 games = 1MB, which is a lot of games per MB.
We'll run out of CPU before memory at that rate.